Extract tint color correction logic
