+++ Wire-iOS/Sources/UserInterface/Conversation Options/TintCorrectedActivityController.swift
@@ -20,16 +20,15 @@ import UIKit
 
 // Subclass intended to work around https://stackoverflow.com/questions/25795065/ios-8-uiactivityviewcontroller-and-uialertcontroller-button-text-color-uses-wind
 final class TintCorrectedActivityViewController: UIActivityViewController {
-    private var windowTintColor: UIColor?
-
+    private var overrider = TintColorOverrider()
+    
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
-        windowTintColor = UIApplication.shared.delegate?.window??.tintColor
-        UIApplication.shared.delegate?.window??.tintColor = .wr_color(fromColorScheme: ColorSchemeColorTextForeground, variant: .light)
+        overrider.override()
     }
     
     override func viewWillDisappear(_ animated: Bool) {
         super.viewWillDisappear(animated)
-        UIApplication.shared.delegate?.window??.tintColor = windowTintColor
+        overrider.restore()
     }
 }
